+++ era-compiler-solidity/src/zksolc/arguments.rs
@@ -69,6 +69,8 @@ pub struct Arguments {
     pub fallback_to_optimizing_for_size: bool,
 
     /// Pass arbitary space-separated options to LLVM.
+    /// The argument must be a single quoted string following a `=` separator.
+    /// Example: `--llvm-options='-eravm-jump-table-density-threshold=10'`.
     #[structopt(long = "llvm-options")]
     pub llvm_options: Option<String>,
 
@@ -78,9 +80,10 @@ pub struct Arguments {
     #[structopt(long = "disable-solc-optimizer")]
     pub disable_solc_optimizer: bool,
 
-    /// Specify the path to the `solc` executable. By default, the one in `${PATH}` is used.
-    /// Yul mode: `solc` is used for source code validation, as `zksolc` itself assumes that the input Yul is valid.
-    /// LLVM IR mode: `solc` is unused.
+    /// Specify the path to a `solc` executable.
+    /// Solidity mode: if not provided, `solc` is also searched in `${PATH}`.
+    /// Yul mode: `solc` is optional for additional Yul validation, as `zksolc` has limited Yul verification capabilities.
+    /// LLVM IR and EraVM assembly modes: `solc` is unused.
     #[structopt(long = "solc")]
     pub solc: Option<String>,